+++ Shared/PerformanceCounters/Interface/ICounters.cs
@@ -8,6 +8,6 @@ namespace Shared.PerformanceCounters.Interface
         int ConnnectedPlayers { get; set; }
         int CPU { get; set; }
         int MaxTickTimeInMs { get; set; }
-        int Memory { get; set; }
+        decimal Memory { get; set; }
     }
 }
